requirements.txt
aiofiles==23.1.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
APScheduler==4.0.0a4
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configobj==5.0.8
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.1
firebase-admin==6.2.0
frozenlist==1.4.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
h11==0.14.0
htmldom==2.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
msgpack==1.0.7
multidict==6.0.4
openpyxl==3.1.2
orjson==3.9.10
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
sqlalchemy-json==0.7.0
starlette==0.27.0
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
xlrd==2.0.1
yarl==1.9.2
